[FLINK-20877][table-runtime-blink] Refactor BytesHashMap and BytesMultiMap to support window key
1. Make BytesMap decouple with BinaryRowData key type
2. Extract pages operations from AbstractRowDataSerializer into PagedTypeSerializer
3. Introduce WindowKey structure and WindowKeySerializer
4. Introduce BytesWindowHashMap and BytesWindowMultiMap to support window key
5. Refactor tests to also cover the new introduced BytesWindowHashMap and BytesWindowMultiMap
