+++ flink-table/flink-table-runtime-blink/src/test/java/org/apache/flink/table/runtime/util/collections/binary/BytesWindowMultiMapTest.java
@@ -20,27 +20,36 @@ package org.apache.flink.table.runtime.util.collections.binary;
 
 import org.apache.flink.runtime.memory.MemoryManager;
 import org.apache.flink.table.data.binary.BinaryRowData;
-import org.apache.flink.table.runtime.typeutils.BinaryRowDataSerializer;
+import org.apache.flink.table.runtime.typeutils.WindowKeySerializer;
+import org.apache.flink.table.runtime.util.WindowKey;
 import org.apache.flink.table.types.logical.LogicalType;
 
-/** Verify the correctness of {@link BytesMultiMap}. */
-public class BytesMultiMapTest extends BytesMultiMapTestBase<BinaryRowData> {
+import java.util.Random;
 
-    public BytesMultiMapTest() {
-        super(new BinaryRowDataSerializer(KEY_TYPES.length));
+/** Verify the correctness of {@link BytesWindowMultiMap}. */
+public class BytesWindowMultiMapTest extends BytesMultiMapTestBase<WindowKey> {
+
+    public BytesWindowMultiMapTest() {
+        super(new WindowKeySerializer(KEY_TYPES.length));
     }
 
     @Override
-    public BytesMultiMapBase<BinaryRowData> createBytesMultiMap(
+    public BytesMultiMapBase<WindowKey> createBytesMultiMap(
             MemoryManager memoryManager,
             int memorySize,
             LogicalType[] keyTypes,
             LogicalType[] valueTypes) {
-        return new BytesMultiMap(this, memoryManager, memorySize, keyTypes, valueTypes);
+        return new BytesWindowMultiMap(this, memoryManager, memorySize, keyTypes, valueTypes);
     }
 
     @Override
-    public BinaryRowData[] generateRandomKeys(int num) {
-        return getRandomizedInputs(num);
+    public WindowKey[] generateRandomKeys(int num) {
+        final Random rnd = new Random(RANDOM_SEED);
+        BinaryRowData[] keys = getRandomizedInputs(num, rnd, true);
+        WindowKey[] windowKeys = new WindowKey[num];
+        for (int i = 0; i < num; i++) {
+            windowKeys[i] = new WindowKey(rnd.nextLong(), keys[i]);
+        }
+        return windowKeys;
     }
 }
